requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
brotli==1.1.0
pytz==2023.3